tdf#109126 fix parsing color tint in BE files

+++ src/lib/PMDTypes.h
@@ -89,7 +89,7 @@ struct PMDCharProperties
   uint16_t m_superSubSize;
   uint16_t m_superPos;
   uint16_t m_subPos;
-  uint8_t m_tint;
+  uint16_t m_tint;
 
   PMDCharProperties(const uint16_t length, const uint16_t fontFace, const uint16_t fontSize, const uint8_t fontColor, const uint8_t boldItalicUnderline, const uint8_t superSubscript, const int16_t kerning, const uint16_t superSubSize, const uint16_t superPos, const uint16_t subPos, const uint16_t tint)
     : m_length(length), m_fontFace(fontFace), m_fontSize(fontSize), m_fontColor(fontColor), m_boldItalicUnderline(boldItalicUnderline), m_superSubscript(superSubscript), m_kerning(kerning), m_superSubSize(superSubSize), m_superPos(superPos), m_subPos(subPos), m_tint(tint)
